Several components were refactored. Updated README file.
