+++ client/src/app/components/Pagination/Pagination.jsx
@@ -1,12 +1,8 @@
+import { usePagination } from '../../hooks/usePagination';
 import styles from './Pagination.module.css';
 
 export const Pagination = ({ currentPage, totalPages, onPageChange }) => {
-    // const pageNumbers = Array.from({ length: totalPages }, (_, i) => i + 1);
-    const pageNumbers = [];
-    for (let i = 1; i <= totalPages; i++) {
-        pageNumbers.push(i);
-    }
-
+    const pageNumbers = usePagination(totalPages);
 
     return (
         <div className={styles.pagination}>
