--- client/src/app/helpers/index.js
@@ -1,3 +1,0 @@
-export const capitalizeFirstLetter = word => {
-    return word.charAt(0).toUpperCase() + word.slice(1);
-};
