+++ client/src/app/utils/index.js
@@ -1,5 +1,12 @@
 import { orderCards, filterCards, cleanFilters } from "../redux/actions";
 
+export const capitalizeFirstLetter = (word) => {
+    if (typeof word === 'string' && word.length > 0) {
+        return word.charAt(0).toUpperCase() + word.slice(1);
+    }
+    return '';
+};
+
 export const sortPokemons = (list, order) => {
     if (order === 'all') {
         return [...list];
